requirements.txt
numpy
opencv-python
PyYAML
pytest
pyautogui
PyGetWindow
pywin32
playwright
